pyproject.toml
[project]
name = "pycmsauto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "pywin32>=308",
    "uiautomator2>=3.2.8",
    "websocket>=0.2.1",
]
